Prevent loading of unused backends
This change changes how the backends are imported,
instead of importing all of them every backend is
loaded lazy only when its needed.

This solves the problem that dependencies from unused
backends need to be installed even if the are not
used (e.g. symspellpy depends on editdistpy which
needs to be build from source on Windows requiring a
compiler ...).