feat: Add guards to disable checking for sections
In some instances users want to disable spell checking for sections of a
document, instead of enabling/disabling it for the whole document. This
introduces spell checking guards in the form of special HTML comments.
For everything enclosed in a
<!-- mkdocs-spellcheck-disable --> ... <!-- mkdocs-spellcheck-enable -->
block, no spell checking is performed.

Issue #33: https://github.com/pawamoy/mkdocs-spellcheck/issues/33